Install path in revserse order.
Previously l2_multi installs all the rules for all the switches
in the path regardless of their order.
This update gaurantees that rules are installed in the reverse order
of the path, so when the packet is pushed back to the network, it's
gauratneed that all the rules are installed.